// packages/sdk/src/types/generated/Order.ts
// Typelizer digest 39b608bc4333e1f86bb7c3bf988c6d69
//
// DO NOT MODIFY: This file was automatically generated by Typelizer.
import type {Discount, LineItem, Fulfillment, Payment, Address, GiftCard, Market} from '@/types'

interface Order {
  id: string;
  market_id: string | null;
  channel_id: string | null;
  number: string;
  email: string;
  customer_note: string | null;
  currency: string;
  locale: string | null;
  total_quantity: number;
  fulfillment_status: string | null;
  payment_status: string | null;
  completed_at: string | null;
  item_total: string | null;
  display_item_total: string | null;
  adjustment_total: string | null;
  display_adjustment_total: string | null;
  discount_total: string | null;
  display_discount_total: string | null;
  tax_total: string | null;
  display_tax_total: string | null;
  included_tax_total: string | null;
  display_included_tax_total: string | null;
  additional_tax_total: string | null;
  display_additional_tax_total: string | null;
  total: string | null;
  display_total: string | null;
  gift_card_total: string | null;
  display_gift_card_total: string | null;
  amount_due: string | null;
  display_amount_due: string | null;
  delivery_total: string | null;
  display_delivery_total: string | null;
  store_credit_total: string | null;
  display_store_credit_total: string | null;
  covered_by_store_credit: boolean;
  discounts: Array<Discount>;
  items: Array<LineItem>;
  fulfillments: Array<Fulfillment>;
  payments: Array<Payment>;
  billing_address: Address | null;
  shipping_address: Address | null;
  gift_card: GiftCard | null;
  market: Market | null;
}

export default Order;