// packages/sdk/src/types/generated/PasswordResetRequestedEvent.ts
// Typelizer digest a23ed5997cf23d88a529180c5ff7c52b
//
// DO NOT MODIFY: This file was automatically generated by Typelizer.

interface PasswordResetRequestedEvent {
  id: string;
  email: string;
  reset_token: string;
  store_id: string | null;
  redirect_url?: string;
}

export default PasswordResetRequestedEvent;